Signals Library
This Library is being created as a part of a redesign. The redesign of
the application to utilize class objects will make construction of
application easier in terms of working with external packages and
libraries. Custom classes inherit from standard library classes
and a few pip installed packages. I plan to package this library up a
little bit because the tooling may be useful in other projects that want
to automate certain backend tasks and give a few developer
friendly UX upgrades when working with threads for new programmers.