Counters module added to Signals Library
Counter module creates counter classes that are used by the threader
classes.The Counter class is an Abstract Base Class interface for
counter implementation and can not be used directly. If you create a
counter that inherits that class you have to implement(overwrite) all the
methods decorated with abstractmethod before it'll run.The Counter class is a
type like integers or booleans.